Move topic card into design module
