+++ feature/topics/src/main/kotlin/uk/govuk/app/topics/ui/widget/TopicsWidget.kt
@@ -1,39 +1,22 @@
 package uk.govuk.app.topics.ui.widget
 
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Star
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.OutlinedCard
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextLayoutResult
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import uk.govuk.app.design.ui.component.BodyBoldLabel
 import uk.govuk.app.design.ui.component.MediumVerticalSpacer
 import uk.govuk.app.design.ui.component.SmallHorizontalSpacer
 import uk.govuk.app.design.ui.component.SmallVerticalSpacer
 import uk.govuk.app.design.ui.component.Title3BoldLabel
+import uk.govuk.app.design.ui.component.TopicCard
 import uk.govuk.app.design.ui.theme.GovUkTheme
 import uk.govuk.app.topics.TopicsViewModel
 import uk.govuk.app.topics.data.remote.model.TopicItem
@@ -87,9 +70,11 @@ private fun TopicsGrid(
 
                         val topicIndex = (rowIndex * columnCount) + columnIndex
                         if (topicIndex < topicsCount) {
-                            TopicsCard(
-                                topic = topics[topicIndex],
-                                modifier = Modifier.weight(1f)
+                            TopicCard(
+                                title = topics[topicIndex].title,
+                                modifier = Modifier
+                                    .fillMaxHeight()
+                                    .weight(1f)
                             )
                         } else {
                             Box(Modifier.weight(1f)) {  }
@@ -101,59 +86,6 @@ private fun TopicsGrid(
                     }
                 }
                 MediumVerticalSpacer()
-            }
-        }
-    }
-}
-
-@Composable
-private fun TopicsCard(
-    topic: TopicItem,
-    modifier: Modifier = Modifier
-) {
-    OutlinedCard(
-        modifier.fillMaxHeight(),
-        colors = CardDefaults.cardColors(
-            containerColor = GovUkTheme.colourScheme.surfaces.card
-        ),
-        border = BorderStroke(
-            width = 1.dp,
-            color = GovUkTheme.colourScheme.strokes.listDivider
-        )
-    ) {
-        Column(
-            Modifier
-                .padding(GovUkTheme.spacing.medium)
-        ){
-            Icon(
-                imageVector = Icons.Filled.Star,
-                contentDescription = null,
-                modifier = Modifier.size(25.dp),
-                tint = GovUkTheme.colourScheme.surfaces.icon
-            )
-            MediumVerticalSpacer()
-            Spacer(Modifier.weight(1f))
-            Row {
-                var baseline by remember { mutableFloatStateOf(0f) }
-                fun updateBaseline(textLayoutResult: TextLayoutResult) {
-                    baseline = textLayoutResult.size.height - textLayoutResult.lastBaseline
-                }
-                val baselinePadding = with(LocalDensity.current) { baseline.toDp() }
-
-                BodyBoldLabel(
-                    text = topic.title,
-                    modifier = Modifier.weight(1f),
-                    onTextLayout = ::updateBaseline
-                )
-                SmallHorizontalSpacer()
-                Icon(
-                    painterResource(uk.govuk.app.design.R.drawable.ic_chevron),
-                    contentDescription = null,
-                    modifier = Modifier
-                        .align(Alignment.Bottom)
-                        .padding(bottom = baselinePadding),
-                    tint = GovUkTheme.colourScheme.textAndIcons.trailingIcon
-                )
             }
         }
     }
